requirements.txt
python-nmap==0.7.1
scikit-learn==1.4.2
numpy==1.26.4
rich==13.7.1
jinja2==3.1.4
